refactor: change code structure and add more comments
