+++ contracts/adapters/swap/pryzm/src/lib.rs
@@ -1,6 +1,7 @@
 pub mod contract;
 pub mod error;
 pub mod state;
-pub mod swap;
+pub mod execution;
 pub mod reply_id;
 pub mod consts;
+pub mod simulate;
