+++ contracts/adapters/swap/pryzm/src/state.rs
@@ -1,7 +1,7 @@
 use std::collections::VecDeque;
 use cosmwasm_std::Addr;
 use cw_storage_plus::Item;
-use crate::swap::SwapExecutionStep;
+use crate::execution::SwapExecutionStep;
 
 pub const ENTRY_POINT_CONTRACT_ADDRESS: Item<Addr> = Item::new("entry_point_contract_address");
